# backend/docs/swagger.yaml
components:
  schemas:
...
    AggregationBody:
      properties:
        field:
          type: string
        size:
          type: integer
        sqon:
          $ref: '#/components/schemas/Sqon'
      type: object
    ApiError:
      description: ApiError represents an error
      properties:
        detail: {}
        message:
          type: string
        status:
          type: integer
      required:
      - message
      - status
      type: object
    AutoCompleteTerm:
      properties:
        highlight:
          $ref: '#/components/schemas/Term'
        source:
          $ref: '#/components/schemas/Term'
      type: object
...
    InterpretationGermline:
      properties:
        classification:
          type: string
        classification_criterias:
          items:
            type: string
          type: array
          uniqueItems: false
        condition:
          type: string
        created_at:
          type: string
        created_by:
          type: string
        created_by_name:
          type: string
        id:
          type: string
        interpretation:
          type: string
        locus_id:
          type: string
        metadata:
          $ref: '#/components/schemas/InterpretationMetadata'
        pubmed:
          items:
            $ref: '#/components/schemas/InterpretationPubmed'
          type: array
          uniqueItems: false
        sequencing_id:
          type: string
        transcript_id:
          type: string
        transmission_modes:
          items:
            type: string
          type: array
          uniqueItems: false
        updated_at:
          type: string
        updated_by:
          type: string
        updated_by_name:
          type: string
      type: object
...
    PubmedCitation:
      properties:
        id:
          type: string
        nlm:
          $ref: '#/components/schemas/PubmedCitationDetails'
      type: object
...
  securitySchemes:
    bearerauth:
      bearerFormat: JWT
      scheme: bearer
      type: http
...
info:
  description: This is the API for Radiant data platform.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Radiant API
  version: "1.0"
openapi: 3.1.0
paths:
  /hpo/autocomplete:
    get:
      description: Retrieve AutoCompleteTerm list of HPO terms matching input string
        with highlighted
      operationId: hpoTermAutoComplete
      parameters:
      - description: Prefix
        in: query
        name: prefix
        required: true
        schema:
          type: string
      - description: Limit
        in: query
        name: limit
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/AutoCompleteTerm'
                type: array
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Internal Server Error
      security:
      - bearerauth: []
      summary: Get AutoCompleteTerm list of matching input string with highlighted
      tags:
      - hpo
  /interpretations/germline:
    get:
      description: Search interpretation germline
      operationId: SearchInterpretationGermline
      responses:
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Internal Server Error
      security:
      - bearerauth: []
      summary: Search interpretation germline
      tags:
      - interpretations
  /interpretations/germline/{sequencing_id}/{locus_id}/{transcript_id}:
    get:
      description: Get interpretation germline
      operationId: GetInterpretationGermline
      parameters:
      - description: Sequencing ID
        in: path
        name: sequencing_id
        required: true
        schema:
          type: string
      - description: Locus ID
        in: path
        name: locus_id
        required: true
        schema:
          type: string
      - description: Transcript ID
        in: path
        name: transcript_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterpretationGermline'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Not Found
      security:
      - bearerauth: []
      summary: Get interpretation germline
      tags:
      - interpretations
    post:
      description: Create or Update interpretation germline
      operationId: PostInterpretationGermline
      parameters:
      - description: Sequencing ID
        in: path
        name: sequencing_id
        required: true
        schema:
          type: string
      - description: Locus ID
        in: path
        name: locus_id
        required: true
        schema:
          type: string
      - description: Transcript ID
        in: path
        name: transcript_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InterpretationGermline'
        description: Interpretation Body
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterpretationGermline'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Bad Request
      security:
      - bearerauth: []
      summary: Create or Update interpretation germline
      tags:
      - interpretations
  /interpretations/pubmed/{citation_id}:
    get:
      description: Get pubmed citation by ID
      operationId: GetPubmedCitation
      parameters:
      - description: Citation ID
        in: path
        name: citation_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PubmedCitation'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Not Found
      security:
      - bearerauth: []
      summary: Get pubmed citation by ID
      tags:
      - interpretations
  /interpretations/somatic:
    get:
      description: Search interpretation somatic
      operationId: SearchInterpretationSomatic
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/InterpretationSomatic'
                type: array
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Internal Server Error
      security:
      - bearerauth: []
      summary: Search interpretation somatic
      tags:
      - interpretations
  /interpretations/somatic/{sequencing_id}/{locus_id}/{transcript_id}:
    get:
      description: Get interpretation somatic
      operationId: GetInterpretationSomatic
      parameters:
      - description: Sequencing ID
        in: path
        name: sequencing_id
        required: true
        schema:
          type: string
      - description: Locus ID
        in: path
        name: locus_id
        required: true
        schema:
          type: string
      - description: Transcript ID
        in: path
        name: transcript_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterpretationSomatic'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Not Found
      security:
      - bearerauth: []
      summary: Get interpretation somatic
      tags:
      - interpretations
    post:
      description: Create or Update interpretation somatic
      operationId: PostInterpretationSomatic
      parameters:
      - description: Sequencing ID
        in: path
        name: sequencing_id
        required: true
        schema:
          type: string
      - description: Locus ID
        in: path
        name: locus_id
        required: true
        schema:
          type: string
      - description: Transcript ID
        in: path
        name: transcript_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InterpretationSomatic'
        description: Interpretation Body
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterpretationSomatic'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Bad Request
      security:
      - bearerauth: []
      summary: Create or Update interpretation somatic
      tags:
      - interpretations
  /mondo/autocomplete:
    get:
      description: Retrieve AutoCompleteTerm list of mondo terms matching input string
        with highlighted
      operationId: mondoTermAutoComplete
      parameters:
      - description: Prefix
        in: query
        name: prefix
        required: true
        schema:
          type: string
      - description: Limit
        in: query
        name: limit
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/AutoCompleteTerm'
                type: array
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Internal Server Error
      security:
      - bearerauth: []
      summary: Get AutoCompleteTerm list of matching input string with highlighted
      tags:
      - mondo
  /occurrences/{seq_id}/{locus_id}/expended:
    get:
      description: Retrieve ExpendedOccurrence data for a given locus ID
      operationId: getExpendedOccurrence
      parameters:
      - description: Sequence ID
        in: path
        name: seq_id
        required: true
        schema:
          type: string
      - description: Locus ID
        in: path
        name: locus_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpendedOccurrence'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Not Found
      security:
      - bearerauth: []
      summary: Get a ExpendedOccurrence
      tags:
      - occurrences
  /occurrences/{seq_id}/aggregate:
    post:
      description: Aggregate occurrences for a given sequence ID
      operationId: aggregateOccurrences
      parameters:
      - description: Sequence ID
        in: path
        name: seq_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AggregationBody'
        description: Aggregation Body
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Aggregation'
                type: array
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Bad Request
      security:
      - bearerauth: []
      summary: Aggregate occurrences
      tags:
      - occurrences
  /occurrences/{seq_id}/count:
    post:
      description: Counts occurrences for a given sequence ID
      operationId: countOccurrences
      parameters:
      - description: Sequence ID
        in: path
        name: seq_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CountBody'
        description: Count Body
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Count'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Bad Request
      security:
      - bearerauth: []
      summary: Count occurrences
      tags:
      - occurrences
...
  /sequencing/{seq_id}:
    get:
      description: Retrieve Sequencing data for a given sequence ID
      operationId: getSequencing
      parameters:
      - description: Sequence ID
        in: path
        name: seq_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sequencing'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Not Found
      security:
      - bearerauth: []
      summary: Get a Sequencing
      tags:
      - sequencing
...
servers:
- url: /